[refactor] Make taichi/common self contained (#2200)
* refactor taichi/common

* set -fPIC

* use singleton pattern for Logger

* reorganize platform macros

* Update taichi/common/logging.cpp


* apply suggestions
