+++ taichi/python/export_misc.cpp
@@ -6,6 +6,7 @@
 #include "taichi/backends/metal/api.h"
 #include "taichi/backends/opengl/opengl_api.h"
 #include "taichi/common/core.h"
+#include "taichi/common/interface.h"
 #include "taichi/common/task.h"
 #include "taichi/math/math.h"
 #include "taichi/python/exception.h"
@@ -115,10 +116,12 @@ void export_misc(py::module &m) {
       .def("test", &Benchmark::test)
       .def("initialize", &Benchmark::initialize);
 
-#define TI_EXPORT_LOGGING(X) \
-  m.def(#X, [](const std::string &msg) { taichi::logger.X(msg); });
+#define TI_EXPORT_LOGGING(X)               \
+  m.def(#X, [](const std::string &msg) {   \
+    taichi::Logger::get_instance().X(msg); \
+  });
 
-  m.def("flush_log", []() { taichi::logger.flush(); });
+  m.def("flush_log", []() { taichi::Logger::get_instance().flush(); });
 
   TI_EXPORT_LOGGING(trace);
   TI_EXPORT_LOGGING(debug);
@@ -131,12 +134,14 @@ void export_misc(py::module &m) {
 
   m.def("print_all_units", print_all_units);
   m.def("set_core_state_python_imported", CoreState::set_python_imported);
-  m.def("set_logging_level",
-        [](const std::string &level) { logger.set_level(level); });
-  m.def("logging_effective", [](const std::string &level) {
-    return logger.is_level_effective(level);
+  m.def("set_logging_level", [](const std::string &level) {
+    Logger::get_instance().set_level(level);
   });
-  m.def("set_logging_level_default", []() { logger.set_level_default(); });
+  m.def("logging_effective", [](const std::string &level) {
+    return Logger::get_instance().is_level_effective(level);
+  });
+  m.def("set_logging_level_default",
+        []() { Logger::get_instance().set_level_default(); });
   m.def("set_core_trigger_gdb_when_crash",
         CoreState::set_trigger_gdb_when_crash);
   m.def("test_raise_error", test_raise_error);
