+++ taichi/system/traceback.cpp
@@ -3,7 +3,8 @@
     The use of this software is governed by the LICENSE file.
 *******************************************************************************/
 
-#include "taichi/common/core.h"
+#include "taichi/system/traceback.h"
+
 #include <vector>
 #include <iostream>
 #include <string>
