+++ taichi/system/traceback.h
@@ -1,0 +1,6 @@
+#include "taichi/common/core.h"
+
+namespace taichi {
+
+TI_EXPORT void print_traceback();
+}
